+++ projects/angular-auth-oidc-client/src/lib/flows/flows.service.ts
@@ -174,8 +174,9 @@ export class FlowsService {
 
     // STEP 2 Refresh Token
     private refreshTokensRequestTokens(
-            callbackContext: CallbackContext,
-            customParams?: { [key: string]: string | number | boolean }): Observable<CallbackContext> {
+        callbackContext: CallbackContext,
+        customParams?: { [key: string]: string | number | boolean }
+    ): Observable<CallbackContext> {
         let headers: HttpHeaders = new HttpHeaders();
         headers = headers.set('Content-Type', 'application/x-www-form-urlencoded');