+++ projects/angular-auth-oidc-client/src/lib/oidc.security.service.ts
@@ -21,7 +21,6 @@ import { TokenHelperService } from './utils/tokenHelper/oidc-token-helper.servic
 
 @Injectable()
 export class OidcSecurityService {
-
     get configuration(): PublicConfiguration {
         return {
             configuration: this.configurationProvider.openIDConfiguration,
@@ -158,6 +157,10 @@ export class OidcSecurityService {
 
     // Code Flow with PCKE or Implicit Flow
     authorize(authOptions?: AuthOptions) {
+        if (authOptions?.customParams) {
+            this.storagePersistanceService.write('storageCustomRequestParams', authOptions?.customParams);
+        }
+
         this.loginService.login(authOptions);
     }
 
@@ -178,6 +181,10 @@ export class OidcSecurityService {
     }
 
     forceRefreshSession(customParams?: { [key: string]: string | number | boolean }) {
+        if (customParams) {
+            this.storagePersistanceService.write('storageCustomRequestParams', customParams);
+        }
+
         return this.refreshSessionService.forceRefreshSession(customParams);
     }